Variablen ausgetaucht durch objekte
